+++ java/com/turbovnc/rfb/Options.java
@@ -1,7 +1,7 @@
 /* Copyright (C) 2002-2005 RealVNC Ltd.  All Rights Reserved.
  * Copyright (C) 2010 TigerVNC Team
  * Copyright (C) 2011 Brian P. Hinz
- * Copyright (C) 2012-2013, 2015-2018, 2020-2021 D. R. Commander.
+ * Copyright (C) 2012-2013, 2015-2018, 2020-2022 D. R. Commander.
  *                                               All Rights Reserved.
  *
  * This is free software; you can redistribute it and/or modify
@@ -403,6 +403,10 @@ public class Options {
 
   public boolean isSecTypeSupported(int secType) {
     Iterator<Integer> i;
+
+    if (sessMgrActive && Params.sessMgrAuto.getValue() &&
+        secType == RFB.SECTYPE_VNCAUTH)
+      return true;
 
     for (i = enabledSecTypes.iterator(); i.hasNext();)
       if ((Integer)i.next() == secType)